Improved error handling

+++ Matrix Playground/Pages/Chat/ChatController.swift
@@ -45,6 +45,8 @@ struct ChatController: View {
     /// Deletes a chat from CoreData store
     func stopChat() {
         do {
+            guard self.sessionData.mxRestClient != nil else {throw ChatError.notLoggedIn}
+            guard self.sessionData.mxRestClient!.credentials.userId != nil else {throw ChatError.notLoggedIn}
             try self.messagingLogic.deleteChat(
                 chat: chatDetails,
                 ownerUserId: self.sessionData.mxRestClient!.credentials.userId!,
@@ -62,6 +64,8 @@ struct ChatController: View {
     /// Stops the user transmitting thier location to a user, and deltes the chat if they are not receiving location updates
     func stopTransmission() {
         do {
+            guard self.sessionData.mxRestClient != nil else {throw ChatError.notLoggedIn}
+            guard self.sessionData.mxRestClient!.credentials.userId != nil else {throw ChatError.notLoggedIn}
             // If we're not receiving and don't wish to send any more hen just delete the chat
             if (chatDetails.receiving == false) { self.stopChat() }
             chatDetails.sending = false
@@ -80,6 +84,8 @@ struct ChatController: View {
     /// Begins transmission of location to a remote user
     func startTransmission() {
         do {
+            guard self.sessionData.mxRestClient != nil else {throw ChatError.notLoggedIn}
+            guard self.sessionData.mxRestClient!.credentials.userId != nil else {throw ChatError.notLoggedIn}
             chatDetails.sending = true
             try messagingLogic.updateChat(
                 chat: chatDetails,
@@ -103,6 +109,9 @@ struct ChatController: View {
     /// handleOnAppear
     /// Handles synchronisation when the page is loaded
     func handleOnAppear() {
+        guard self.sessionData.mxRestClient != nil else {return}
+        guard self.sessionData.mxRestClient!.credentials.userId != nil else {return}
+        
         let userFetchRequest = NSFetchRequest<NSManagedObject>(entityName: "UserDetails")
         userFetchRequest.predicate = NSPredicate(format: "userId == %@", self.sessionData.mxRestClient!.credentials.userId!)
         var userFetchedResults = [UserDetails]()
